Improve pixmap cache to not depend on file name.
Previously, the cache used file name to refer to pixmap data.
While this is valid approach for most of use cases, it's not ideal
for this application because the entire point of it is to rename files.

In order not to depend on file name we query the filesystem for unique
file id (aka inode), which we combine with device id and last write time
to make sure we reload images that have been modified externally.

Sadly, Qt does not provide inode or device id in QFile API, so we have to
implement it in platform-dependent way.